requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.69.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-anthropic==0.3.9
langchain-community==0.3.19
langchain-core==0.3.41
langchain-google-genai==2.0.11
langchain-groq==0.2.4
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langchain-xai==0.2.1
langsmith==0.3.11
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.0.2
openai==1.65.4
orjson==3.10.15
packaging==24.2
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
